Tidy fixtures
- Remove unused fixtures
- Move local fixtures to the relevant test files
- Use tmp_path fixture to create and move test files
