+++ tests/utils.py
@@ -8,7 +8,8 @@ TEST_BUCKET_PARQUET_FILEPATH_WITH_SHORTNAME = "gs://ssb-staging-dapla-felles-dat
 
 TEST_RESOURCES_DIRECTORY = Path("tests/resources")
 
-TEST_PARQUET_FILEPATH = TEST_RESOURCES_DIRECTORY / "person_data_v1.parquet"
+TEST_PARQUET_FILE_NAME = "person_data_v1.parquet"
+TEST_PARQUET_FILEPATH = TEST_RESOURCES_DIRECTORY / TEST_PARQUET_FILE_NAME
 TEST_SAS7BDAT_FILEPATH = TEST_RESOURCES_DIRECTORY / "sasdata.sas7bdat"
 TEST_PARQUET_GZIP_FILEPATH = TEST_RESOURCES_DIRECTORY / "person_data_v1.parquet.gzip"
 
@@ -16,9 +17,6 @@ TEST_EXISTING_METADATA_DIRECTORY = TEST_RESOURCES_DIRECTORY / "existing_metadata
 TEST_EXISTING_METADATA_FILE_NAME = "person_data_v1__DOC.json"
 TEST_EXISTING_METADATA_FILEPATH = (
     TEST_EXISTING_METADATA_DIRECTORY / TEST_EXISTING_METADATA_FILE_NAME
-)
-TEST_RESOURCES_METADATA_DOCUMENT = (
-    TEST_RESOURCES_DIRECTORY / TEST_EXISTING_METADATA_FILE_NAME
 )
 
 TEST_EXISTING_METADATA_WITH_VALID_ID_DIRECTORY = (
